requirements.txt
google-cloud-firestore>=2.11.1
google-cloud-secret-manager>=2.16.0
google-generativeai>=0.3.0
twilio>=8.0.0
flask>=3.0.0
functions-framework>=3.0.0
python-dotenv>=1.0.0
